Lowered default highWaterMark to 20 for objectMode: true
In object mode, highWaterMark is the maximum number of objects to keep in the
stream's internal buffer. For long stream pipelines, keeping 1000 objects in
the buffers of both the readable and writable sides of each transform stream
can end up using a lot of memory and causing your program to hang.